Samle logikk for å trigge satsjustering i SatsjusteringService.kt
